==> 0.0_DATASET_masters_check_populations.ipynb <==
## **0.0_DATASET_masters_check_populations**

### GOALS of this script:
* check the populations contained in the data files
* check numbers and assignment of the samples to the populations
* all markers in the Unterseer vcf files only have one alt allele
* the variants/FILTER_PASS column has no meaning 
* strand column in the hapmap file has no meaning
* the variants/LID and variants/ID columns contain the same thing 
* the variants/VRT and variants/is_snp columns contain the same thing 
* checked if read_vcf() and vcf_to_dataframe() create the same dataframe

### *Import packages*

import numpy as np
import seaborn as sns
sns.set_style('white')
sns.set_style('ticks')
import pandas as pd
import matplotlib.pyplot as plt
%matplotlib inline
import allel
import random
import h5py
import bcolz
import csv
from itertools import chain
import os
import re
from collections import Counter
from matplotlib.patches import Patch
from matplotlib.lines import Line2D
from matplotlib.legend_handler import HandlerLine2D

xxxx

### *Check the lines in Unterseer_2016_landraces*

#### *Read in hapmap file*

unterseer_2016_landraces_hapmap=pd.read_csv("/Users/kschul38/Documents/tsinfer-project/data/2_processed/europe_maize_dataset_600k/unterseer_2016_landraces_hapmap.hmp.txt",sep="\t")
unterseer_2016_landraces_hapmap

**Check the number of samples**

#get the samples in the hapmap
unterseer_2016_landraces_hapmap_columns=list(unterseer_2016_landraces_hapmap.columns)
unterseer_2016_landraces_hapmap_columns=unterseer_2016_landraces_hapmap_columns[11:]
#unterseer_2016_landraces_hapmap_columns
print(len(unterseer_2016_landraces_hapmap_columns))

**Check the assignment to the populations**

#unterseer_2016_landraces_hapmap_columns

#Counter([elem.split(".", 1)[0] for elem in unterseer_2016_landraces_hapmap_columns])

#### *Read in vcf file*

unterseer_2016_landraces_vcf= allel.read_vcf('/Users/kschul38/Documents/tsinfer-project/data/1_raw/5_Unterseer_2016/TUM-PLANTBREEDING_Maize600k_landraces.vcf', fields='*')

sorted(unterseer_2016_landraces_vcf.keys())

Counter(unterseer_2016_landraces_vcf['variants/is_snp'])

Counter(unterseer_2016_landraces_vcf['variants/VRT']==unterseer_2016_landraces_vcf['variants/is_snp'])

Counter(unterseer_2016_landraces_vcf['variants/LID']==unterseer_2016_landraces_vcf['variants/ID'])

Counter(unterseer_2016_landraces_vcf['variants/numalt'])

Counter(unterseer_2016_landraces_vcf['variants/FILTER_PASS'])

#unterseer_2016_landraces_vcf['variants/QUAL']

**Check the number of samples**

#get the samples in the .vcf
unterseer_2016_landraces_vcf_samples= unterseer_2016_landraces_vcf["samples"]
print(len(unterseer_2016_landraces_vcf_samples))

**Check the assignment to the populations**

#Counter([elem.split(".", 1)[0] for elem in unterseer_2016_landraces_vcf])

### *Check the lines in Unterseer_2016_elite lines*

unterseer_2016_elite_hapmap=pd.read_csv("/Users/kschul38/Documents/tsinfer-project/data/2_processed/europe_maize_dataset_600k/unterseer_2016_elite_hapmap.hmp.txt",sep="\t")
unterseer_2016_elite_hapmap

**Check the number of samples**

#get the samples in the .vcf
unterseer_2016_elite_hapmap_columns=list(unterseer_2016_elite_hapmap.columns)
unterseer_2016_elite_hapmap_columns=unterseer_2016_elite_hapmap_columns[11:]
#unterseer_2016_elite_hapmap_columns
print(len(unterseer_2016_elite_hapmap_columns))

#unterseer_2016_elite_hapmap_columns

#### *Read in vcf file*

unterseer_2016_elite_vcf= allel.read_vcf('/Users/kschul38/Documents/tsinfer-project/data/1_raw/5_Unterseer_2016/TUM-PLANTBREEDING_Maize600k_elitelines.vcf',fields='*')

sorted(unterseer_2016_elite_vcf.keys())

Counter(unterseer_2016_elite_vcf['variants/VRT']==unterseer_2016_elite_vcf['variants/is_snp'])

Counter(unterseer_2016_elite_vcf['variants/LID']==unterseer_2016_elite_vcf['variants/ID'])

Counter(unterseer_2016_elite_vcf['variants/numalt'])

Counter(unterseer_2016_elite_vcf['variants/FILTER_PASS'])

**Check the number of samples**

#get the samples in the .vcf
unterseer_2016_elite_vcf_samples= unterseer_2016_elite_vcf["samples"]
print(len(unterseer_2016_elite_vcf_samples))

#unterseer_2016_elite_vcf

### *Check the lines in Mayer_2017_landraces*

#read in the .txt file from the /data/1_raw/4_Mayer_2017 folder 
mayer_2017_landraces=pd.read_csv('/Users/kschul38/Documents/tsinfer-project/data/1_raw/4_Mayer_2017/TUM-PLANTBREEDING_Maize600k_35landraces_952individuals_raw.txt', sep=" ")
mayer_2017_landraces=mayer_2017_landraces.reset_index()
mayer_2017_landraces
#markers already used as row names 

**Check the number of samples**

##get the samples in the hapmap
mayer_2017_landraces_columns=list(mayer_2017_landraces.columns)
mayer_2017_landraces_columns=mayer_2017_landraces_columns[6:]
#mayer_2017_landraces_columns
print(len(mayer_2017_landraces_columns))

**Check the assignment to the populations**

#Counter([elem.split(".", 1)[0] for elem in mayer_2017_landraces_columns])

### *Check the lines in Mayer_2020_dhlines*

#read in the .txt file from the /data/1_raw/4_Mayer_2017 folder 
mayer_2020_dhlines=pd.read_csv('/Users/kschul38/Documents/tsinfer-project/data/1_raw/3_Mayer_2020/Mayer_et_al_genotypes_DHlines_600k_raw.txt.gz', sep=" ")
#markers already used as row names 

mayer_2020_dhlines

**Check the number of samples**

##get the samples in the hapmap
mayer_2020_dhlines_columns=list(mayer_2020_dhlines.columns)
mayer_2020_dhlines_columns=mayer_2020_dhlines_columns[4:]
#mayer_2020_dhlines_columns
print(len(mayer_2020_dhlines_columns))

**Check the assignment to the populations**

pattern = '[0-9]'
Counter([re.sub(pattern, '', i) for i in mayer_2020_dhlines_columns])

### *Check the strand in Unterseer Hapmap*

Counter(unterseer_2016_landraces_hapmap["strand"])

Counter(unterseer_2016_elite_hapmap["strand"])

### *Check if read_vcf and vcf_to_dataframe create the same dataframe*

**Unterseer_landraces**

unterseer_landraces_dataframe = allel.vcf_to_dataframe('/Users/kschul38/Documents/tsinfer-project/data/1_raw/5_Unterseer_2016/TUM-PLANTBREEDING_Maize600k_landraces.vcf', fields= ['variants/ID','variants/AD', 'variants/CMT'])
unterseer_landraces_dataframe = unterseer_landraces_dataframe.reset_index(drop=True)

unterseer_2016_landraces_vcf_dataframe = unterseer_landraces_dataframe.rename(columns={"AD": "variants/AD_2016_landraces","CMT": "variants/CMT_2016_landraces", "ID": "rs#_2016_landraces" })

unterseer_2016_landraces_vcf_dataframe

unterseer_2016_landraces_vcf_3= allel.read_vcf('/Users/kschul38/Documents/tsinfer-project/data/1_raw/5_Unterseer_2016/TUM-PLANTBREEDING_Maize600k_landraces.vcf',fields=['variants/ID','variants/AD', 'variants/CMT'])

unterseer_2016_landraces_vcf_file = pd.DataFrame(data=unterseer_2016_landraces_vcf_3)
#unterseer_2016_landraces_vcf = unterseer_2016_landraces_vcf.set_index("variants/ID")
unterseer_2016_landraces_vcf_file = unterseer_2016_landraces_vcf_file.rename(columns={"variants/AD": "variants/AD_2016_landraces","variants/CMT": "variants/CMT_2016_landraces", "variants/ID": "rs#_2016_landraces" })

unterseer_2016_landraces_vcf_file=unterseer_2016_landraces_vcf_file[['rs#_2016_landraces','variants/AD_2016_landraces','variants/CMT_2016_landraces']]

unterseer_2016_landraces_vcf_file.equals(unterseer_2016_landraces_vcf_dataframe)

**Unterseer_elite**

unterseer_elite_dataframe = allel.vcf_to_dataframe('/Users/kschul38/Documents/tsinfer-project/data/1_raw/5_Unterseer_2016/TUM-PLANTBREEDING_Maize600k_elitelines.vcf',fields=['variants/ID','variants/AD', 'variants/CMT'])
unterseer_elite_dataframe = unterseer_elite_dataframe.reset_index(drop=True)

unterseer_2016_elite_vcf_dataframe = unterseer_elite_dataframe.rename(columns={"AD": "variants/AD_2016_elite","CMT": "variants/CMT_2016_elite", "ID": "rs#_2016_elite"})

unterseer_2016_elite_vcf_dataframe

unterseer_2016_elite_vcf_3= allel.read_vcf('/Users/kschul38/Documents/tsinfer-project/data/1_raw/5_Unterseer_2016/TUM-PLANTBREEDING_Maize600k_elitelines.vcf',fields=['variants/ID','variants/AD', 'variants/CMT'])

unterseer_2016_elite_vcf_file = pd.DataFrame(data=unterseer_2016_elite_vcf_3)
#unterseer_2016_elite_vcf = unterseer_2016_elite_vcf.set_index("variants/ID")
unterseer_2016_elite_vcf_file = unterseer_2016_elite_vcf_file.rename(columns={"variants/AD": "variants/AD_2016_elite","variants/CMT": "variants/CMT_2016_elite", "variants/ID": "rs#_2016_elite"})

unterseer_2016_elite_vcf_file = unterseer_2016_elite_vcf_file[['rs#_2016_elite','variants/AD_2016_elite','variants/CMT_2016_elite']]

unterseer_2016_elite_vcf_file

unterseer_2016_elite_vcf_file.equals(unterseer_2016_elite_vcf_dataframe)

xxxxx

###  *Check if final_columns_xxx.csv & final_columns_xxx_for_masters.csv are the same*

**final_columns_xxx.csv**

final_columns_unterseer_2016_elite=pd.read_csv('/Users/kschul38/Documents/tsinfer-project/data/2_processed/europe_maize_dataset_600k/final_columns_unterseer_2016_elite.csv', sep="\t")

final_columns_unterseer_2016_landraces=pd.read_csv('/Users/kschul38/Documents/tsinfer-project/data/2_processed/europe_maize_dataset_600k/final_columns_unterseer_2016_landraces.csv', sep="\t")

final_columns_mayer_2017_landraces=pd.read_csv('/Users/kschul38/Documents/tsinfer-project/data/2_processed/europe_maize_dataset_600k/final_columns_mayer_2017_landraces.csv', sep="\t")

final_columns_mayer_2020_dhlines=pd.read_csv('/Users/kschul38/Documents/tsinfer-project/data/2_processed/europe_maize_dataset_600k/final_columns_mayer_2020_dhlines.csv', sep="\t")

**final_columns_xxx_for_masters.csv**

final_columns_unterseer_2016_elite_for_masters=pd.read_csv('final_columns_unterseer_2016_elite_for_masters.csv', sep="\t")

final_columns_unterseer_2016_landraces_for_masters=pd.read_csv('final_columns_unterseer_2016_landraces_for_masters.csv', sep="\t")

final_columns_mayer_2017_landraces_for_masters=pd.read_csv('final_columns_mayer_2017_landraces_for_masters.csv', sep="\t")

final_columns_mayer_2020_dhlines_for_masters=pd.read_csv('final_columns_mayer_2020_dhlines_for_masters.csv', sep="\t")

**Compare**

final_columns_unterseer_2016_elite.equals(final_columns_unterseer_2016_elite_for_masters)

final_columns_unterseer_2016_landraces.equals(final_columns_unterseer_2016_landraces_for_masters)

final_columns_mayer_2017_landraces.equals(final_columns_mayer_2017_landraces_for_masters)

final_columns_mayer_2020_dhlines.equals(final_columns_mayer_2020_dhlines_for_masters)

###  *Check if filtered_xxx.csv & filtered_xxx_for_masters.csv are the same*

**filtered_xxx.csv**

filtered_unterseer_2016_elite_right_pos=pd.read_csv("/Users/kschul38/Documents/tsinfer-project/data/2_processed/europe_maize_dataset_600k/filtered_unterseer_2016_elite.csv", sep="\t")

filtered_unterseer_2016_landraces_right_pos=pd.read_csv("/Users/kschul38/Documents/tsinfer-project/data/2_processed/europe_maize_dataset_600k/filtered_unterseer_2016_landraces.csv", sep="\t")

filtered_mayer_2017_landraces_suffix=pd.read_csv("/Users/kschul38/Documents/tsinfer-project/data/2_processed/europe_maize_dataset_600k/filtered_mayer_2017_landraces.csv",sep="\t")

filtered_mayer_2020_dhlines_suffix=pd.read_csv("/Users/kschul38/Documents/tsinfer-project/data/2_processed/europe_maize_dataset_600k/filtered_mayer_2020_dhlines.csv",sep="\t")

**filtered_xxx_for_masters.csv**

filtered_unterseer_2016_elite_for_masters=pd.read_csv("filtered_unterseer_2016_elite_for_masters.csv",sep="\t")

filtered_unterseer_2016_landraces_for_masters=pd.read_csv("filtered_unterseer_2016_landraces_for_masters.csv",sep="\t")

filtered_mayer_2017_landraces_for_masters=pd.read_csv("filtered_mayer_2017_landraces_for_masters.csv",sep="\t")

filtered_mayer_2020_dhlines_for_masters=pd.read_csv("filtered_mayer_2020_dhlines_for_masters.csv",sep="\t")

**Compare**

filtered_unterseer_2016_elite_right_pos.equals(filtered_unterseer_2016_elite_for_masters)

filtered_unterseer_2016_landraces_right_pos.equals(filtered_unterseer_2016_landraces_for_masters)

filtered_mayer_2017_landraces_suffix.equals(filtered_mayer_2017_landraces_for_masters)

filtered_mayer_2020_dhlines_suffix.equals(filtered_mayer_2020_dhlines_for_masters)

###  *Check if alleles are the same in vcf & array_600k_info*

**Array 600k info**

array_600k_info=pd.read_csv("/Users/kschul38/Documents/tsinfer-project/data/2_processed/europe_maize_dataset_600k/600k_array_info_header_removed.txt",sep="\t")

array_600k_info

array_600k_info.loc[array_600k_info['ID'] == "AX-90525992"]

**Final vcf file**

callset = allel.vcf_to_dataframe('/Users/kschul38/Documents/tsinfer-project/data/2_processed/europe_maize_dataset_600k/merge_merge_all_chrom_all_edits_with_allel_column_vcf.vcf')

#callset

callset.loc[callset['ID'] == "AX-90525992"]

**One of the input datasets**

mayer_2017_landraces.loc[mayer_2017_landraces["index"]=="AX-90525992"]

Counter(mayer_2017_landraces.loc[387028])

### **Check if is_snp markers were removed**

*Unterseer*

unterseer_2016_landraces_vcf_test= allel.vcf_to_dataframe('/Users/kschul38/Documents/tsinfer-project/data/1_raw/5_Unterseer_2016/TUM-PLANTBREEDING_Maize600k_landraces.vcf', fields='*')

#unterseer_2016_landraces_vcf_test

is_snp_variants=unterseer_2016_landraces_vcf_test[~unterseer_2016_landraces_vcf['variants/is_snp']]

#is_snp_variants

Counter(test["is_snp"])

is_snp_variants_AD=is_snp_variants["AD"]

#is_snp_variants_AD

*Array 600k info*

#get all the genotype-calls from the array.txt
genotype_call=array_600k_info["alleles_f"].values.ravel()

#check that number of genotype calls is correct
len(genotype_call)

#get a list of all unique calls
unique_values_array = pd.unique(genotype_call)

#unique_values_array

#look at complete list
#unique_values_array

#remove the "typical genotype calls" - so that just indels remain
unique_values_array = np.delete(unique_values_array, [0,1,2,3,4,5])

#check that it only contains a-typical genotype calls
#unique_values_array

#create a boolean over the whole allele column, where typical calls == True & atypical calls == False 
array_marker_filter_table=~array_600k_info["alleles_f"].isin(unique_values_array)

array_marker_filter_table

#get all marker IDs from the array.txt
array_non_typical_genotype_calls_marker=array_600k_info['ID']

#list of all marker IDs
#array_non_typical_genotype_calls_marker

#convert to array
array_non_typical_genotype_calls_marker=array_non_typical_genotype_calls_marker.to_numpy()

#array of all marker IDs
#array_non_typical_genotype_calls_marker

#filter marker array using the inversion of the boolean (TRUE=atypical, FALSE=typical)
array_non_typical_genotype_calls_marker=array_non_typical_genotype_calls_marker[~array_marker_filter_table]

#list of marker IDs that belong to the a-typical genotype calls
array_non_typical_genotype_calls_marker

#check length array 600k info
len(array_non_typical_genotype_calls_marker)

#check length unterseer
len(is_snp_variants_AD)

*Check the intersection*

len(list(set(array_non_typical_genotype_calls_marker) & set(unterseer_test)))

len(set(array_non_typical_genotype_calls_marker).intersection(unterseer_test))

len(np.intersect1d(array_non_typical_genotype_calls_marker,unterseer_test))

len(list(set(array_non_typical_genotype_calls_marker) - set(unterseer_test)))

### **Check if vcf info columns are the same**

callset_new = allel.vcf_to_dataframe('merge_merge_all_chrom_all_edits_with_allel_column_for_masters_vcf.vcf')

callset_new

callset_old = allel.vcf_to_dataframe('/Users/kschul38/Documents/tsinfer-project/data/2_processed/europe_maize_dataset_600k/merge_merge_all_chrom_all_edits_with_allel_column_vcf.vcf')

callset_old

callset_new.equals(callset_old)

### **Check if intermediate merge scripts are the same**

merge_merge_turn_around_new=pd.read_csv("merge_merge_all_chromosomes_turn_around_for_masters.hmp.txt", sep="\t")

merge_merge_turn_around_new_str=merge_merge_turn_around_new.astype(str)

merge_merge_turn_around_old=pd.read_csv("/Users/kschul38/Documents/tsinfer-project/data/2_processed/europe_maize_dataset_600k/merge_merge_all_chromosomes_turn_around.hmp.txt", sep="\t")

merge_merge_turn_around_old_str=merge_merge_turn_around_old.astype(str)

merge_merge_turn_around_new_str.equals(merge_merge_turn_around_old_str)

### **Check if the final allele column and the array info allele column are the same**

#read-in the array.txt
check_strand=pd.read_csv("/Users/kschul38/Documents/tsinfer-project/data/2_processed/europe_maize_dataset_600k/600k_array_info_header_removed.txt", sep="\t", dtype = str)
check_strand

#get all the genotype-calls from the array.txt
genotype_call=check_strand["alleles_f"].values.ravel()

#check that number of genotype calls is correct
len(genotype_call)

#get a list of all unique calls
unique_values_array = pd.unique(genotype_call)

#look at complete list
#unique_values_array

#remove the "typical genotype calls" - so that just indels remain
unique_values_array = np.delete(unique_values_array, [0,1,2,3,4,5])

#check that it only contains a-typical genotype calls
#unique_values_array

#create a boolean over the whole allele column, where typical calls == True & atypical calls == False 
array_marker_filter_table=~check_strand["alleles_f"].isin(unique_values_array)

array_marker_filter_table

**Get the filtered markers from the array.txt**

filter_boolean_array_txt=array_marker_filter_table #taken from the non-typical genotypes removed to use here
#Counter(filter_boolean_array_txt)

#filter_boolean_array_txt

Counter(filter_boolean_array_txt)

#check_strand[["ID","alleles_f"]]
array_txt_alleles=check_strand[["ID","alleles_f"]]
array_txt_alleles_merge=array_txt_alleles.rename(columns={"ID":"index_test"}) #rename the column for merge 
array_txt_alleles_merge_filter=array_txt_alleles_merge[filter_boolean_array_txt] # remove the lines with ne non-typical genotype calls 

array_txt_alleles_merge_filter

Counter(array_txt_alleles_merge_filter["alleles_f"])

205822-205756

203955-203870

57702-57654

57514-57468

35367-35345

49082-49079

66+85+48+46+22+3

Counter(allel_column)
Counter({'T/C': 205756,
         'G/A': 203870,
         'G/C': 49079,
         'T/A': 35345,
         'T/G': 57468,
         'A/C': 57654,
         'A': 56,
         'T': 59,
         'C': 66,
         'G': 89})

56+59+66+89

no_header=pd.read_csv("no_header.vcf", sep="\t")

no_header

vcf_check=no_header.drop(['1','0'], axis=1)

### *Check the genotype call in the .vcf file*

#recheck the genotype calls for everything but the first 11 columns
check_vcf_geno=vcf_check.iloc[:,7:]

check_vcf_geno

check_genotype_counts_values_vcf=check_vcf_geno.values.ravel()

flatten_list_unique_values_vcf=pd.unique(check_genotype_counts_values_vcf)

flatten_list_unique_values_vcf
