==> 0.0_INPUT_masters_make_metadata_file.ipynb <==
## **0.0_INPUT_master_make_metadata_file**

### GOALS of this script:
* make metadatafiles 
* same as the ones used just without obs_het column
* hex_col colors put in by hand

#import all needed datasets
import pandas as pd
import numpy as np
import scipy as sp

#### *Read in the Mayer_2017 metadata excel file*

meta_mayer_2017_landraces= pd.read_excel(r'/Users/kschul38/Documents/tsinfer-project/data/1_raw/4_Mayer_2017/mayer_2017_landraces_meta.xlsx', sheet_name='Table 1')

meta_mayer_2017_landraces

meta_for_plotly=meta_mayer_2017_landraces

#remove last line that sums the total number of samples
meta_for_plotly=meta_for_plotly.drop(35)

#### *Drop irrelevant columns*

meta_for_plotly=meta_for_plotly.drop(['Regiona', 'Country','Seed Source'], axis=1)

#meta_for_plotly

#### *Add columns for marker color, shape & size*

#add columns for marker color, shape & size
meta_for_plotly['hex_col']= " "
meta_for_plotly['shape']= "circle"
meta_for_plotly['size']= 5

#### *Duplicate name column*

meta_for_plotly['Full name duplicate'] = meta_for_plotly['Full name']

#### *Add PCA_shape columns*

meta_for_plotly['shape_PCA']= "X"

meta_for_plotly

#### *Add the lines from the Unterseer_landraces by hand - long und lat found via google maps*

unterseer_additional_landraces = pd.DataFrame([["Altreier", "AL",46.277,11.365,24,"","circle",5,"Altreier","X"],
                                    ["Fleimstal","FL",46.303,11.600,24,"","circle",5,"Fleimstal","X"],
                                    ["Sornay","SO",46.627, 5.178,23,"","circle",5,"Sornay","X"]],columns=list('ABCDEFGHJK'))

unterseer_additional_landraces

unterseer_additional_landraces=unterseer_additional_landraces.rename(columns={"A": "Full name", "B": "Abb.", "C":"Latitudeb","D":"Longitudeb","E":"Num of Individuals","F":"hex_col","G":"shape","H":"size","J":"Full name duplicate","K":"shape_PCA"})


unterseer_additional_landraces

meta_for_plotly=meta_for_plotly.append(unterseer_additional_landraces)

meta_for_plotly

#reindex the table
meta_for_plotly=meta_for_plotly.reset_index(drop=True)
meta_for_plotly

#### *Edit the landraces corresponding to the Mayer_2020 DH lines rows and add them*

KL_row=meta_for_plotly.iloc[[8]]
KL_row["shape_PCA"]="^"

PE_row=meta_for_plotly.iloc[[21]]
PE_row["shape_PCA"]="^"

LL_row=meta_for_plotly.iloc[[12]]
LL_row["shape_PCA"]="^"

#drop the columns
meta_for_plotly=meta_for_plotly.drop([8, 21,12])

#append the rows at the end
meta_for_plotly=meta_for_plotly.append(KL_row, ignore_index=False)
meta_for_plotly=meta_for_plotly.append(PE_row, ignore_index=False)
meta_for_plotly=meta_for_plotly.append(LL_row, ignore_index=False)

meta_for_plotly

#reindex the table
meta_for_plotly=meta_for_plotly.reset_index(drop=True)
#meta_for_plotly

#### *Add the three DH lines from the Mayer_2020 paper by hand*

#set up the columns
meta_mayer_2020_dh_lines = pd.DataFrame([["Kemater Landmais Gelb DH", "KL DH",47.255,11.274,501,"","square",10,"Kemater Landmais Gelb DH","s"],
                                         ["Petkuser Ferdinand Rot DH","PE DH",51.981,13.349,409,"","square",10,"Petkuser Ferdinand Rot DH","s"],
                                         ["Lalin DH", "LL DH",42.661,-8.111,31,"","square",10,"Lalin DH","s"]],columns=list('ABCDEFGHJK'))


#meta_mayer_2020_dh_lines

#rename the columns
meta_mayer_2020_dh_lines=meta_mayer_2020_dh_lines.rename(columns={"A": "Full name", "B": "Abb.", "C":"Latitudeb","D":"Longitudeb","E":"Num of Individuals","F":"hex_col","G":"shape","H":"size","J":"Full name duplicate","K":"shape_PCA"})


#meta_mayer_2020_dh_lines

#### *Append the Mayer_2020 table to the meta_for_plotly table*

meta_for_plotly=meta_for_plotly.append(meta_mayer_2020_dh_lines)

#meta_for_plotly

#reindex the table
meta_for_plotly=meta_for_plotly.reset_index(drop=True)
meta_for_plotly

#### *Append Unterseer_2016 elite lines info* & make the two metadata file

**Make meta_full_landraces_with_PCA_shapes_reordered_new_colorpalette**

meta_for_plotly_1=meta_for_plotly

#set up the columns
meta_unterseer_2016_1 = pd.DataFrame([["Unterseer_2016_elite flint", "flint",0.0,0.0,66,"","diamond",10,"Unterseer_2016_elite flint","p"],
                                    ["Unterseer_2016_elite dent","dent",0.0,0.0,77,"","diamond",10,"Unterseer_2016_elite dent","p"],
                                    ["Unterseer_2016_elite other","-",0.0,0.0,19,"","diamond",10,"Unterseer_2016_elite other","p"]],columns=list('ABCDEFGHJK'))


#rename the columns
meta_unterseer_2016_1=meta_unterseer_2016_1.rename(columns={"A": "Full name", "B": "Abb.", "C":"Latitudeb","D":"Longitudeb","E":"Num of Individuals","F":"hex_col","G":"shape","H":"size","J":"Full name duplicate","K":"shape_PCA"})


#meta_unterseer_2016_1

meta_for_plotly_1=meta_for_plotly.append(meta_unterseer_2016_1, ignore_index=False)

#reindex the table
meta_for_plotly_1=meta_for_plotly_1.reset_index(drop=True)
meta_for_plotly_1

**Write to csv file**

#meta_for_plotly_1.to_csv('meta_full_landraces_with_PCA_shapes_reordered_new_colorpalette.csv', index=False) 

**Make metadata_for_PCA**

meta_for_plotly_2=meta_for_plotly

#set up the columns
meta_unterseer_2016_2 = pd.DataFrame([["Unterseer_2016_elite flint", "flint",0.0,0.0,66,"","diamond",10,"european breeding lines","p"],
                                      ["Unterseer_2016_elite dent","dent",0.0,0.0,77,"","diamond",10,"european breeding lines","p"]],columns=list('ABCDEFGHJK'))


#rename the columns
meta_unterseer_2016_2=meta_unterseer_2016_2.rename(columns={"A": "Full name", "B": "Abb.", "C":"Latitudeb","D":"Longitudeb","E":"Num of Individuals","F":"hex_col","G":"shape","H":"size","J":"Full name duplicate","K":"shape_PCA"})


meta_unterseer_2016_2

meta_for_plotly_2=meta_for_plotly.append(meta_unterseer_2016_2, ignore_index=False)

#reindex the table
meta_for_plotly_2=meta_for_plotly_2.reset_index(drop=True)
meta_for_plotly_2

**Write to csv file**

#meta_for_plotly_2.to_csv('metadata_for_PCA.csv', index=False)
