==> 1.1_DATASET_masters_convert_Unterseer_vcf_to_hmp.txt_tassel_5.0.ipynb <==
## **DATASET_masters_1.1.convert_Unterseer_vcf_to_hmp.txt_tassel_5.0.ipynb**

### GOALS of this script:
* convert the Unterseer_2016 VCF (.vcf) files into Hapmap files (.hmp.txt)

## **NOTES:**

* open Tassel GUI: ./start\textunderscore tassel.pl -Xmx4g
* load file 
* save as .hmp.txt
