==> 1.8_DATASET_masters_compare_allel_columnwise.ipynb <==
## **1.8_DATASET_masters_compare_allel_columnwise.ipynb**

### GOALS of this script:
* part three of whole merging workflow for the masters
* numbering in accordance to workflow_europe_maize_dataset_prep_masters
* turn around calls for the array.txt column changed from mask to replace 
* combination/check of:
    * merging_script_final-Copy1.ipynb
    * merging_script_final.ipynb
    * the corresponding number steps
* OUT: xxx_allel_col_for_masters.csv (optional)

### *Import packages*

import gzip
import allel
import pandas as pd
import numpy as np
import tskit
import tsinfer
import sys
import json
import csv
from IPython.display import SVG
from IPython.display import HTML
from collections import Counter
import matplotlib.pyplot as plt
import seaborn as sns
from openpyxl import Workbook
from progressbar import ProgressBar
sns.set_style('white')
sns.set_style('ticks')

## *Import files*

filtered_unterseer_2016_elite=pd.read_csv("filtered_unterseer_2016_elite_for_masters.csv", sep="\t")

filtered_unterseer_2016_elite

filtered_unterseer_2016_landraces=pd.read_csv("filtered_unterseer_2016_landraces_for_masters.csv", sep="\t")

filtered_unterseer_2016_landraces

filtered_mayer_2017_landraces=pd.read_csv("filtered_mayer_2017_landraces_for_masters.csv", sep="\t")

filtered_mayer_2017_landraces

filtered_mayer_2020_dhlines=pd.read_csv("filtered_mayer_2020_dhlines_for_masters.csv", sep="\t")

filtered_mayer_2020_dhlines

## **1.8 Compare allel columns between all datasets**

### *1.8.1 Turn around the double genotype calls for the Unterseer datasets*

#### *Unterseer_2016_elite*

#get the alleles that correspond to the marker
unterseer_2016_elite_allel_comp=filtered_unterseer_2016_elite.loc[:,["variants/AD_2016_elite","alleles_2016_elite"]]
unterseer_2016_elite_allel_comp

**Turn around genotype calls**

*Look at the current genotype calls*

unterseer_2016_elite_allel_test= unterseer_2016_elite_allel_comp["alleles_2016_elite"].values.ravel()

unterseer_2016_elite_unique_alleles=pd.unique(unterseer_2016_elite_allel_test)

unterseer_2016_elite_unique_alleles

#number of different genotype calls 
len(unterseer_2016_elite_unique_alleles)

*Exchange C/T for T/C*

#update_TC=unterseer_2016_elite_allel_comp.alleles_2016_elite.mask(unterseer_2016_elite_allel_comp.alleles_2016_elite == "C/T","T/C")

unterseer_2016_elite_allel_comp=unterseer_2016_elite_allel_comp.replace("C/T","T/C")

#unterseer_2016_elite_allel_comp.update(update_TC)

#check if CTs were replace with TCs
#unterseer_2016_elite_allel_comp
#pd.unique(unterseer_2016_elite_allel_comp.alleles_2016_elite.values.ravel())
len(pd.unique(unterseer_2016_elite_allel_comp.alleles_2016_elite.values.ravel())) #number reduced by one? - yes!

*Exchange G/T for T/G* 

#update_TG=unterseer_2016_elite_allel_comp.alleles_2016_elite.mask(unterseer_2016_elite_allel_comp.alleles_2016_elite == "G/T","T/G")

unterseer_2016_elite_allel_comp=unterseer_2016_elite_allel_comp.replace("G/T","T/G")

#unterseer_2016_elite_allel_comp.update(update_TG)

#check if GTs were replace with TGs
#unterseer_2016_elite_allel_comp
#pd.unique(unterseer_2016_elite_allel_comp.alleles_2016_elite.values.ravel())
len(pd.unique(unterseer_2016_elite_allel_comp.alleles_2016_elite.values.ravel())) #number reduced by one? - yes!

*Exchange A/T for T/A*

#update_TA=unterseer_2016_elite_allel_comp.alleles_2016_elite.mask(unterseer_2016_elite_allel_comp.alleles_2016_elite == "A/T","T/A")

unterseer_2016_elite_allel_comp=unterseer_2016_elite_allel_comp.replace("A/T","T/A")

#unterseer_2016_elite_allel_comp.update(update_TA)

#check if ATs were replace with TAs
#unterseer_2016_elite_allel_comp
#pd.unique(unterseer_2016_elite_allel_comp.alleles_2016_elite.values.ravel())
len(pd.unique(unterseer_2016_elite_allel_comp.alleles_2016_elite.values.ravel())) #number reduced by one? - yes!

*Exchange A/G for G/A*

#update_GA=unterseer_2016_elite_allel_comp.alleles_2016_elite.mask(unterseer_2016_elite_allel_comp.alleles_2016_elite == "A/G","G/A")

unterseer_2016_elite_allel_comp=unterseer_2016_elite_allel_comp.replace("A/G","G/A")

#unterseer_2016_elite_allel_comp.update(update_GA)

#check if AGs were replace with GAs
#unterseer_2016_elite_allel_comp
#pd.unique(unterseer_2016_elite_allel_comp.alleles_2016_elite.values.ravel())
len(pd.unique(unterseer_2016_elite_allel_comp.alleles_2016_elite.values.ravel())) #number reduced by one? - yes!

*Exchange C/G for G/C*

#update_GC=unterseer_2016_elite_allel_comp.alleles_2016_elite.mask(unterseer_2016_elite_allel_comp.alleles_2016_elite == "C/G","G/C")

unterseer_2016_elite_allel_comp=unterseer_2016_elite_allel_comp.replace("C/G","G/C")

#unterseer_2016_elite_allel_comp.update(update_GC)

#check if CGs were replace with GCs
#unterseer_2016_elite_allel_comp
#pd.unique(unterseer_2016_elite_allel_comp.alleles_2016_elite.values.ravel())
len(pd.unique(unterseer_2016_elite_allel_comp.alleles_2016_elite.values.ravel())) #number reduced by one? - yes!

*Exchange C/A for A/C*

#update_AC=unterseer_2016_elite_allel_comp.alleles_2016_elite.mask(unterseer_2016_elite_allel_comp.alleles_2016_elite == "C/A","A/C")

unterseer_2016_elite_allel_comp=unterseer_2016_elite_allel_comp.replace("C/A","A/C")

#unterseer_2016_elite_allel_comp.update(update_AC)

#check if CAs were replace with ACs
#unterseer_2016_elite_allel_comp
#pd.unique(unterseer_2016_elite_allel_comp.alleles_2016_elite.values.ravel())
len(pd.unique(unterseer_2016_elite_allel_comp.alleles_2016_elite.values.ravel())) #number reduced by one? - yes!

pd.unique(unterseer_2016_elite_allel_comp.alleles_2016_elite.values.ravel()) #number reduced by one? - yes!

unterseer_2016_elite_allel_comp

#### *Unterseer_2016_landraces*

#get the alleles that corespond to the marker 
unterseer_2016_landraces_allel_comp=filtered_unterseer_2016_landraces.loc[:,["variants/AD_2016_landraces","alleles_2016_landraces"]]
unterseer_2016_landraces_allel_comp

**Turn around gentoype calls**

*Look at the current genotype calls*

unterseer_2016_landraces_allel_test= unterseer_2016_landraces_allel_comp["alleles_2016_landraces"].values.ravel()

unterseer_2016_landraces_unique_alleles=pd.unique(unterseer_2016_landraces_allel_test)

unterseer_2016_landraces_unique_alleles

#number of different genotype calls 
len(unterseer_2016_landraces_unique_alleles)

*Exchange C/T for T/C*

#update_TC=unterseer_2016_landraces_allel_comp.alleles_2016_landraces.mask(unterseer_2016_landraces_allel_comp.alleles_2016_landraces == "C/T","T/C")

unterseer_2016_landraces_allel_comp=unterseer_2016_landraces_allel_comp.replace("C/T","T/C")

#unterseer_2016_landraces_allel_comp.update(update_TC)

#check if CTs were replace with TCs
#unterseer_2016_landraces_allel_comp
#pd.unique(unterseer_2016_landraces_allel_comp.alleles_2016_landraces.values.ravel())
len(pd.unique(unterseer_2016_landraces_allel_comp.alleles_2016_landraces.values.ravel())) #number reduced by one? - yes!

*Exchange G/T for T/G* 

#update_TG=unterseer_2016_landraces_allel_comp.alleles_2016_landraces.mask(unterseer_2016_landraces_allel_comp.alleles_2016_landraces == "G/T","T/G")

unterseer_2016_landraces_allel_comp=unterseer_2016_landraces_allel_comp.replace("G/T","T/G")

#unterseer_2016_landraces_allel_comp.update(update_TG)

#check if GTs were replace with TGs
#unterseer_2016_landraces_allel_comp
#pd.unique(unterseer_2016_landraces_allel_comp.alleles_2016_landraces.values.ravel())
len(pd.unique(unterseer_2016_landraces_allel_comp.alleles_2016_landraces.values.ravel())) #number reduced by one? - yes!

*Exchange A/T for T/A*

#update_TA=unterseer_2016_landraces_allel_comp.alleles_2016_landraces.mask(unterseer_2016_landraces_allel_comp.alleles_2016_landraces == "A/T","T/A")

unterseer_2016_landraces_allel_comp=unterseer_2016_landraces_allel_comp.replace("A/T","T/A")

#unterseer_2016_landraces_allel_comp.update(update_TA)

#check if ATs were replace with TAs
#unterseer_2016_landraces_allel_comp
#pd.unique(unterseer_2016_landraces_allel_comp.alleles_2016_landraces.values.ravel())
len(pd.unique(unterseer_2016_landraces_allel_comp.alleles_2016_landraces.values.ravel())) #number reduced by one? - yes!

*Exchange A/G for G/A*

#update_GA=unterseer_2016_landraces_allel_comp.alleles_2016_landraces.mask(unterseer_2016_landraces_allel_comp.alleles_2016_landraces == "A/G","G/A")

unterseer_2016_landraces_allel_comp=unterseer_2016_landraces_allel_comp.replace("A/G","G/A")

#unterseer_2016_landraces_allel_comp.update(update_GA)

#check if AGs were replace with GAs
#unterseer_2016_landraces_allel_comp
#pd.unique(unterseer_2016_landraces_allel_comp.alleles_2016_landraces.values.ravel())
len(pd.unique(unterseer_2016_landraces_allel_comp.alleles_2016_landraces.values.ravel())) #number reduced by one? - yes!

*Exchange C/G for G/C*

#update_GC=unterseer_2016_landraces_allel_comp.alleles_2016_landraces.mask(unterseer_2016_landraces_allel_comp.alleles_2016_landraces == "C/G","G/C")

unterseer_2016_landraces_allel_comp=unterseer_2016_landraces_allel_comp.replace("C/G","G/C")

#unterseer_2016_landraces_allel_comp.update(update_GC)

#check if CGs were replace with GCs
#unterseer_2016_landraces_allel_comp
#pd.unique(unterseer_2016_landraces_allel_comp.alleles_2016_landraces.values.ravel())
len(pd.unique(unterseer_2016_landraces_allel_comp.alleles_2016_landraces.values.ravel())) #number reduced by one? - yes!

*Exchange C/A for A/C*

#update_AC=unterseer_2016_landraces_allel_comp.alleles_2016_landraces.mask(unterseer_2016_landraces_allel_comp.alleles_2016_landraces == "C/A","A/C")

unterseer_2016_landraces_allel_comp=unterseer_2016_landraces_allel_comp.replace("C/A","A/C")

#unterseer_2016_landraces_allel_comp.update(update_AC)

#check if CAs were replace with ACs
#unterseer_2016_landraces_allel_comp
#pd.unique(unterseer_2016_landraces_allel_comp.alleles_2016_landraces.values.ravel())
len(pd.unique(unterseer_2016_landraces_allel_comp.alleles_2016_landraces.values.ravel())) #number reduced by one? - yes!

pd.unique(unterseer_2016_landraces_allel_comp.alleles_2016_landraces.values.ravel()) #number reduced by one? - yes!

unterseer_2016_landraces_allel_comp

### *1.8.2 Generate the allele columns for the Mayer datasets*

#### **Mayer_2017**

#get a list of the remaining markers 
mayer_2017_landraces_allel_comp=filtered_mayer_2017_landraces.loc[:,["index_2017_landraces"]]
mayer_2017_landraces_allel_comp

#get the alleles
mayer_2017_landraces_get_alleles=filtered_mayer_2017_landraces.drop(['index_2017_landraces','chr_v4_2017_landraces','pos_v4_2017_landraces','quality_2017_landraces'], axis = 1)

mayer_2017_landraces_get_alleles

#get the remainig genotype calls 
mayer_2017_landraces_get_alleles_str=mayer_2017_landraces_get_alleles.astype(str)

mayer_2017_landraces_get_alleles_str

 pd.unique(mayer_2017_landraces_get_alleles_str.values.ravel())

**Replace the OO with NaN** 

mayer_2017_landraces_get_alleles_reset = mayer_2017_landraces_get_alleles_str.replace(to_replace='OO', value="nan")

mayer_2017_all_geno_values_test_replace =mayer_2017_landraces_get_alleles_reset.values.ravel()

unique_values_mayer_2017_test_replace = pd.unique(mayer_2017_all_geno_values_test_replace)

unique_values_mayer_2017_test_replace

len(unique_values_mayer_2017_test_replace)

**Get the row-wise alleles from the table**

mayer_2017_genotypes_set=list(map(set,mayer_2017_landraces_get_alleles_reset.values))

#list of sets 
mayer_2017_genotypes_set

**Make list of sets into list of lists**

#mayer_2017_genotypes
mayer_2017_genotypes_list=[]

for e in mayer_2017_genotypes_set:
    #print(e)
    geno_row=list(e)
    mayer_2017_genotypes_list.append(geno_row)

mayer_2017_genotypes_list

**Remove "nan" from the rows**

for sublist_2017 in mayer_2017_genotypes_list:
    #print(sublist)
    for element_2017 in sublist_2017:
        #print(element)
        if element_2017 == "nan":
            #print("yes")
            sublist_2017.remove(element_2017)

mayer_2017_genotypes_list

**Merge all the row elements into one**

merged_mayer_2017_genotypes_list=[]

for sublist_without_2017 in mayer_2017_genotypes_list:
    #print(sublist_without_2017)
    s=''.join(sublist_without_2017)
    merged_mayer_2017_genotypes_list.append(s)
    #print(s)

merged_mayer_2017_genotypes_list

substring_T_2017="T"
substring_G_2017="G"
substring_A_2017="A"
substring_C_2017="C"

count_row_2017=[]
all_counts_2017=[]
filter_boolean_2017=[]

**Filter for sites with more than two alleles**

for merged_element in merged_mayer_2017_genotypes_list:
    #print(merged_element)
    
    #count how many times each allel appears in the row 
    count_T_2017=merged_element.count(substring_T_2017)
    count_G_2017=merged_element.count(substring_G_2017)
    count_A_2017=merged_element.count(substring_A_2017)
    count_C_2017=merged_element.count(substring_C_2017)
    #print(count_T)
    
    #add the rowwise counts of the alleles to a list
    count_row_2017=[count_T_2017,count_G_2017,count_A_2017,count_C_2017]
    #add the row counts to a list of lists
    all_counts_2017.append(count_row_2017)

    #count number of non-zero numbers in the count of alleles list (are there more than two alleles )
    num_of_alleles_2017=np.count_nonzero(count_row_2017)
    #print(num_of_alleles)
    
    if num_of_alleles_2017 > 2: #if there are more than two alleles 
        #print(merged_element)
        #print(num_of_alleles)
        filter_boolean_2017.append(False) #append False to the list 
    
    elif num_of_alleles_2017 <=2: #if there are less than two allels
        filter_boolean_2017.append(True)
              
    else:
        print("Problem")
        break  

print(len(filter_boolean_2017))
Counter(filter_boolean_2017)

#print(all_counts_2017[0])
#print(merged_mayer_2017_genotypes_list[0])

**Make the allele column 1**

allel_column_2017=[]

#make the allel column 
for count_row_element_2017 in all_counts_2017 :
            #print(count_row_element)
            if count_row_element_2017[0] != 0 and count_row_element_2017[1] != 0:
                allel_column_2017.extend(["T/G"])
            elif count_row_element_2017[0] != 0 and count_row_element_2017[2] !=0:
                allel_column_2017.extend(["T/A"])
            elif count_row_element_2017[0] != 0 and count_row_element_2017[3] !=0:
                allel_column_2017.extend(["T/C"])
            elif count_row_element_2017[1] != 0 and count_row_element_2017[2] !=0:
                allel_column_2017.extend(["G/A"])
            elif count_row_element_2017[1] != 0 and count_row_element_2017[3] !=0:
                allel_column_2017.extend(["G/C"])
            elif count_row_element_2017[2] != 0 and count_row_element_2017[3] !=0:
                allel_column_2017.extend(["A/C"])
            elif count_row_element_2017[0] != 0 and count_row_element_2017[1] ==0 and count_row_element_2017[2] ==0 and count_row_element_2017[3]==0:
                allel_column_2017.extend(["T"])
            elif count_row_element_2017[1] != 0 and count_row_element_2017[0] ==0 and count_row_element_2017[2] ==0 and count_row_element_2017[3]==0:
                allel_column_2017.extend(["G"])
            elif count_row_element_2017[2] != 0 and count_row_element_2017[0] ==0 and count_row_element_2017[1] ==0 and count_row_element_2017[3]==0:
                allel_column_2017.extend(["A"])
            elif count_row_element_2017[3] != 0 and count_row_element_2017[0] ==0 and count_row_element_2017[1] ==0 and count_row_element_2017[2]==0:
                allel_column_2017.extend(["C"])
            else:
                allel_column_2017.extend(["-/-"])


len(allel_column_2017)

Counter(allel_column_2017)

35180+57495+57340+203402+205273+48989+297+588+570+308

**Make the allele column 2**

merged_condensed_mayer_2017_genotypes_list=[]

for merged_sublist_2017 in merged_mayer_2017_genotypes_list:
    #print(merged_sublist)
    k="".join(set(merged_sublist_2017))
    #print(k)
    merged_condensed_mayer_2017_genotypes_list.append(k)

len(merged_condensed_mayer_2017_genotypes_list)

Counter([len(i) for i in merged_condensed_mayer_2017_genotypes_list])

Counter(merged_condensed_mayer_2017_genotypes_list)

allel_column_2017_2=[]

for condensed_ele in merged_condensed_mayer_2017_genotypes_list:
    if condensed_ele == "AT" or condensed_ele == "TA":
        allel_column_2017_2.append("T/A")
    if condensed_ele == "CA" or condensed_ele == "AC":
        allel_column_2017_2.append("A/C")
    if condensed_ele == "GT" or condensed_ele == "TG":
        allel_column_2017_2.append("T/G")
    if condensed_ele == "GA" or condensed_ele == "AG":
        allel_column_2017_2.append("G/A")
    if condensed_ele == "CT" or condensed_ele == "TC":
        allel_column_2017_2.append("T/C")
    if condensed_ele == "GC" or condensed_ele == "CG":
        allel_column_2017_2.append("G/C")
    if condensed_ele == "T":
        allel_column_2017_2.append("T")
    if condensed_ele == "C":
        allel_column_2017_2.append("C")
    if condensed_ele == "G":
        allel_column_2017_2.append("G")
    if condensed_ele == "A":
        allel_column_2017_2.append("A")

Counter(allel_column_2017_2)

allel_column_2017_2 == allel_column_2017

### **Mayer_2020**

#get a list of the remaining markers 
mayer_2020_dhlines_allel_comp=filtered_mayer_2020_dhlines.loc[:,["marker_2020_dh"]]
mayer_2020_dhlines_allel_comp

#get the alleles
mayer_2020_dhlines_get_alleles=filtered_mayer_2020_dhlines.drop(['marker_2020_dh','chr_2020_dh','pos_2020_dh','quality_2020_dh'], axis = 1)

mayer_2020_dhlines_get_alleles

#get the remainig genotype calls 
mayer_2020_dhlines_get_alleles_str=mayer_2020_dhlines_get_alleles.astype(str)

mayer_2020_dhlines_get_alleles_str

**Replace the OO with NaN** 

mayer_2020_dhlines_get_alleles_reset= mayer_2020_dhlines_get_alleles_str.replace(to_replace='OO', value="nan") 

mayer_2020_all_geno_values_test_replace =mayer_2020_dhlines_get_alleles_reset.values.ravel()

Counter(mayer_2020_dhlines_get_alleles_str.values.ravel())

unique_values_mayer_2020_test_replace = pd.unique(mayer_2020_all_geno_values_test_replace)

unique_values_mayer_2020_test_replace 

len(unique_values_mayer_2020_test_replace)

**Get the row-wise alleles from the table**

mayer_2020_genotypes_set=list(map(set,mayer_2020_dhlines_get_alleles_reset.values))

#list of sets
mayer_2020_genotypes_set

**Make list of sets into list of lists**

#mayer_2020_genotypes
mayer_2020_genotypes_list=[]

for e in mayer_2020_genotypes_set:
    #print(e)
    geno_row=list(e)
    mayer_2020_genotypes_list.append(geno_row)

mayer_2020_genotypes_list

**Remove "nan" from the rows**

for sublist_2020 in mayer_2020_genotypes_list:
    #print(sublist)
    for element_2020 in sublist_2020:
        #print(element)
        if element_2020 == "nan":
            #print("yes")
            sublist_2020.remove(element_2020)

mayer_2020_genotypes_list

**Merge all the row elements into one**

merged_mayer_2020_genotypes_list=[]

for sublist_without_2020 in mayer_2020_genotypes_list:
    #print(sublist_without_2020)
    s=''.join(sublist_without_2020)
    merged_mayer_2020_genotypes_list.append(s)
    #print(s)

merged_mayer_2020_genotypes_list

substring_T_2020="T"
substring_G_2020="G"
substring_A_2020="A"
substring_C_2020="C"

count_row_2020=[]
all_counts_2020=[]
filter_boolean_2020=[]

**Filter for sites with more than two alleles**

for merged_element in merged_mayer_2020_genotypes_list:
    #print(merged_element)
    
    #count how many times each allel appears in the row 
    count_T_2020=merged_element.count(substring_T_2020)
    count_G_2020=merged_element.count(substring_G_2020)
    count_A_2020=merged_element.count(substring_A_2020)
    count_C_2020=merged_element.count(substring_C_2020)
    #print(count_T)
    
    #add the rowwise counts of the alleles to a list
    count_row_2020=[count_T_2020,count_G_2020,count_A_2020,count_C_2020]
    #add the row counts to a list of lists
    all_counts_2020.append(count_row_2020)

    #count number of non-zero numbers in the count of alleles list (are there more than two alleles )
    num_of_alleles_2020=np.count_nonzero(count_row_2020)
    #print(num_of_alleles)
    
    if num_of_alleles_2020 > 2: #if there are more than two alleles 
        #print(merged_element)
        #print(num_of_alleles)
        filter_boolean_2020.append(False) #append False to the list 
    
    elif num_of_alleles_2020 <=2: #if there are less than two allels
        filter_boolean_2020.append(True)
              
    else:
        print("Problem")
        break  

print(len(filter_boolean_2020))
Counter(filter_boolean_2020)

#print(all_counts_2020[0])
#print(merged_mayer_2020_genotypes_list[0])

**Make the allele column**

allel_column_2020=[]

#make the allel column 
for count_row_element_2020 in all_counts_2020 :
            #print(count_row_element)
            if count_row_element_2020[0] != 0 and count_row_element_2020[1] != 0:
                allel_column_2020.extend(["T/G"])
            elif count_row_element_2020[0] != 0 and count_row_element_2020[2] !=0:
                allel_column_2020.extend(["T/A"])
            elif count_row_element_2020[0] != 0 and count_row_element_2020[3] !=0:
                allel_column_2020.extend(["T/C"])
            elif count_row_element_2020[1] != 0 and count_row_element_2020[2] !=0:
                allel_column_2020.extend(["G/A"])
            elif count_row_element_2020[1] != 0 and count_row_element_2020[3] !=0:
                allel_column_2020.extend(["G/C"])
            elif count_row_element_2020[2] != 0 and count_row_element_2020[3] !=0:
                allel_column_2020.extend(["A/C"])
            elif count_row_element_2020[0] != 0 and count_row_element_2020[1] ==0 and count_row_element_2020[2] ==0 and count_row_element_2020[3]==0:
                allel_column_2020.extend(["T"])
            elif count_row_element_2020[1] != 0 and count_row_element_2020[0] ==0 and count_row_element_2020[2] ==0 and count_row_element_2020[3]==0:
                allel_column_2020.extend(["G"])
            elif count_row_element_2020[2] != 0 and count_row_element_2020[0] ==0 and count_row_element_2020[1] ==0 and count_row_element_2020[3]==0:
                allel_column_2020.extend(["A"])
            elif count_row_element_2020[3] != 0 and count_row_element_2020[0] ==0 and count_row_element_2020[1] ==0 and count_row_element_2020[2]==0:
                allel_column_2020.extend(["C"])
            else:
                allel_column_2020.extend(["-/-"])


len(allel_column_2020)

Counter(allel_column_2020)

198429+55506+55623+34412+196435+7788+7645+2991+47580+3033

**Make the allele column 2**

merged_condensed_mayer_2020_genotypes_list=[]

for merged_sublist_2020 in merged_mayer_2020_genotypes_list:
    #print(merged_sublist)
    l="".join(set(merged_sublist_2020))
    #print(k)
    merged_condensed_mayer_2020_genotypes_list.append(l)

len(merged_condensed_mayer_2017_genotypes_list)

Counter([len(i) for i in merged_condensed_mayer_2020_genotypes_list])

all_counts_2020_2=[]

for condensed_ele in merged_condensed_mayer_2020_genotypes_list:
    if condensed_ele == "AT" or condensed_ele == "TA":
        all_counts_2020_2.append("T/A")
    if condensed_ele == "CA" or condensed_ele == "AC":
        all_counts_2020_2.append("A/C")
    if condensed_ele == "GT" or condensed_ele == "TG":
        all_counts_2020_2.append("T/G")
    if condensed_ele == "GA" or condensed_ele == "AG":
        all_counts_2020_2.append("G/A")
    if condensed_ele == "CT" or condensed_ele == "TC":
        all_counts_2020_2.append("T/C")
    if condensed_ele == "GC" or condensed_ele == "CG":
        all_counts_2020_2.append("G/C")
    if condensed_ele == "T":
        all_counts_2020_2.append("T")
    if condensed_ele == "C":
        all_counts_2020_2.append("C")
    if condensed_ele == "G":
        all_counts_2020_2.append("G")
    if condensed_ele == "A":
        all_counts_2020_2.append("A")

Counter(all_counts_2020_2)

Counter(allel_column_2020)

all_counts_2020_2 == allel_column_2020

### *1.8.3 Make allele table*

**Get the filtered markers from the array.txt**

#read-in the array.txt
check_strand=pd.read_csv("/Users/kschul38/Documents/tsinfer-project/data/2_processed/europe_maize_dataset_600k/600k_array_info_header_removed.txt", sep="\t", dtype = str)
check_strand

#get all the genotype-calls from the array.txt
genotype_call=check_strand["alleles_f"].values.ravel()

#check that number of genotype calls is correct
len(genotype_call)

#get a list of all unique calls
unique_values_array = pd.unique(genotype_call)

#look at complete list
#unique_values_array

#remove the "typical genotype calls" - so that just indels remain
unique_values_array = np.delete(unique_values_array, [0,1,2,3,4,5])

#check that it only contains a-typical genotype calls
#unique_values_array

#create a boolean over the whole allele column, where typical calls == True & atypical calls == False 
array_marker_filter_table=~check_strand["alleles_f"].isin(unique_values_array)

#array_marker_filter_table

#Count the number of rows with typical genotype calls TRUE and a-typical genotype calls FALSE
Counter(array_marker_filter_table)

**Add the allel to the marker column for the Mayer datasets**

*Mayer_2020*

print(len(mayer_2020_dhlines_allel_comp))
print(len(allel_column_2020))

mayer_2020_dhlines_allel_comp['allel_column_2020'] = allel_column_2020

mayer_2020_dhlines_allel_comp

*Mayer_2017*

print(len(mayer_2017_landraces_allel_comp))
print(len(allel_column_2017))

mayer_2017_landraces_allel_comp['allel_column_2017'] = allel_column_2017

mayer_2017_landraces_allel_comp

**Rename the marker column for merging**

*Mayer_2017*

#mayer_2017_landraces_allel_comp
mayer_2017_landraces_allel_comp=mayer_2017_landraces_allel_comp.rename(columns={"index_2017_landraces": "index_test"})
mayer_2017_landraces_allel_comp

*Mayer_2020*

#mayer_2020_dhlines_allel_comp
mayer_2020_dhlines_allel_comp=mayer_2020_dhlines_allel_comp.rename(columns={"marker_2020_dh": "index_test"})
mayer_2020_dhlines_allel_comp

*Unterseer_2016_elite*

#unterseer_2016_elite_allel_comp
unterseer_2016_elite_allel_comp=unterseer_2016_elite_allel_comp.rename(columns={"variants/AD_2016_elite": "index_test"})
unterseer_2016_elite_allel_comp

*Unterseer_2016_landraces*

#unterseer_2016_landraces_allel_comp
unterseer_2016_landraces_allel_comp=unterseer_2016_landraces_allel_comp.rename(columns={"variants/AD_2016_landraces": "index_test"})
unterseer_2016_landraces_allel_comp

pd.unique(unterseer_2016_landraces_allel_comp.alleles_2016_landraces)

**Get the filtered markers from the array.txt**

filter_boolean_array_txt=array_marker_filter_table #taken from the non-typical genotypes removed to use here
#Counter(filter_boolean_array_txt)

#filter_boolean_array_txt

Counter(filter_boolean_array_txt)

#check_strand[["ID","alleles_f"]]
array_txt_alleles=check_strand[["ID","alleles_f"]]
array_txt_alleles_merge=array_txt_alleles.rename(columns={"ID":"index_test"}) #rename the column for merge 
array_txt_alleles_merge_filter=array_txt_alleles_merge[filter_boolean_array_txt] # remove the lines with ne non-typical genotype calls 

array_txt_alleles_merge_filter

array_txt_alleles_merge_filter_ravel_unique=array_txt_alleles_merge_filter["alleles_f"].values.ravel()
Counter(array_txt_alleles_merge_filter_ravel_unique) #no untypical genotype calls 

205822+203955+57702+57514+35367+49082

**Check genotype calls & trun around calls for the array.txt column**

*Exchange C/T for T/C*

array_txt_alleles_merge_filter["alleles_f"]=array_txt_alleles_merge_filter["alleles_f"].replace("C/T","T/C")

array_txt_alleles_merge_filter

*Exchange G/T for T/G*

array_txt_alleles_merge_filter["alleles_f"]=array_txt_alleles_merge_filter["alleles_f"].replace("G/T","T/G")

array_txt_alleles_merge_filter

*Exchange A/T and T/A*

array_txt_alleles_merge_filter["alleles_f"]=array_txt_alleles_merge_filter["alleles_f"].replace("A/T","T/A")

*Exchange A/G and G/A*

array_txt_alleles_merge_filter["alleles_f"]=array_txt_alleles_merge_filter["alleles_f"].replace("A/G","G/A")

*Exchange C/G and G/C*

array_txt_alleles_merge_filter["alleles_f"]=array_txt_alleles_merge_filter["alleles_f"].replace("C/G","G/C")

**Check if the lengths of the arrays are ok**

print(len(mayer_2017_landraces_allel_comp))
print(len(mayer_2020_dhlines_allel_comp))
print(len(unterseer_2016_elite_allel_comp))
print(len(unterseer_2016_landraces_allel_comp))
print(len(array_txt_alleles_merge_filter))

**Stepwise merge the columns together**

unterseer_merge=pd.merge(unterseer_2016_landraces_allel_comp,unterseer_2016_elite_allel_comp, on='index_test', how='outer')
unterseer_merge

mayer_merge=pd.merge(mayer_2020_dhlines_allel_comp,mayer_2017_landraces_allel_comp, on='index_test', how='outer')
mayer_merge

merge_merge=pd.merge(unterseer_merge, mayer_merge, on='index_test', how='outer')

merge_merge

**Merge the array.txt column into the table**

merge_merge_array=pd.merge(merge_merge,array_txt_alleles_merge_filter, on='index_test', how='outer')

merge_merge_array

#drop the marker IDs 
merge_merge_geno_comp=merge_merge_array.drop(columns=['index_test'])
merge_merge_geno_comp

#merge_merge_geno_comp.to_csv('merge_merge_geno.csv')

### *1.8.4 Compare alleles row-wise*

**Get a list of the remaining genotypes**

check_merge_merge_geno_comp=merge_merge_geno_comp.astype(str)

pd.unique(check_merge_merge_geno_comp.values.ravel())

**Get the genotype calls**

*Make a list of sets out of it*

merge_merge_geno_comp_str=merge_merge_geno_comp.astype(str)

merge_merge_set=list(map(set,merge_merge_geno_comp_str.values))

#merge_merge_set

#check length of the list of sets 
print(len(merge_merge_set))

*Make a list of lists out of it*

merge_allele_list=[]

for s in merge_merge_set:
    #print(e)
    allele_row=list(s)
    merge_allele_list.append(allele_row)

#check the list of lists 
merge_allele_list

#check length of the list of lists
print(len(merge_allele_list))

**Remove the "nan" form the genotype calls**

for sublist_merge in merge_allele_list:
    #print(sublist_merge)
    for merge_element in sublist_merge:
        #print(merge_element)
        if merge_element == "nan":
            #print("yes")
            sublist_merge.remove(merge_element)

#check if row stays the same 
merge_merge_set[609441]

#check if "nan" call was removed from row - yes!
merge_allele_list[609441]

**Check for the number of alleles per row 1**

allele_check_list=[]

for merge_allele in merge_allele_list:
    #print(merge_allele)
    #print(len(merge_allele))
    if len(merge_allele)==1: #check how many different diff allele calls there are in a row - just one 
        #print("One element" + str(merge_allele))
        for element_allele in merge_allele: #check how long the call is 
            if len(element_allele)==1: #one allele
                #print("Short call: " + element_allele)
                allele_check_list.append(True) #call OK append True
            elif len(element_allele)==3: #two alleles
                #print("Long call: " + element_allele)
                allele_check_list.append(True) #call OK append True
            else: #more than two alleles, problem, break
                #print("More than two alleles" + str(element_allele))
                break
        
    elif len(merge_allele)==2: #check how many different diff allele calls there are - two 
        #print("Two elements" + str(merge_allele))
        for element_allele in merge_allele: #check how long the call is 
            if len(element_allele)==1: #one allele
                subsub=element_allele #is substring 
                #print("Short call: " + subsub)
            elif len(element_allele)==3: #two allele
                fullfull=element_allele #is fullstring 
                #print("Long call: " + fullfull)
            else: #more than two alleles, problem - break 
                #print("More than two alleles" + element_allele)
                break
                
        if subsub in fullfull:#check if the shorter call is the substring of the longer call (it should be A/C with either A or C as second call)
            #print("Found!")
            allele_check_list.append(True) # calls ok append True
        elif subsub not in fullfull: # if substring not in fullstring, problem break
            #print("Not found!" + subsub + " - " + fullfull) #if the small call is not a substring of the large call STOP and print
            break
            
    elif len(merge_allele)==3: #check how many different diff allele calls there are - three 
        print("Three elements" + str(merge_allele))
        allele_check_list.append(True)
        
    elif len(merge_allele)<3: #there should not be more than three different calls per row, problem - break
        print("More than three calls: " + str(merge_allele))
        allele_check_list.append(False)
        break

Counter(allele_check_list)

**Check for the number of alleles per row 2**

merged_condensed_merge_allele_list=[]

for merged_allel in merge_allele_list:
    v="".join(set(merged_allel))
    #print(k)
    merged_condensed_merge_allele_list.append(v)

#merged_condensed_merge_allele_list
len(merged_condensed_merge_allele_list)

merged_second_condensed_merge_allele_list=[]

for sub_merged_allel in merged_condensed_merge_allele_list:
    #print(sub_merged_allel)
    m = sub_merged_allel.replace("/","")
    merged_second_condensed_merge_allele_list.append(m)

#merged_second_condensed_merge_allele_list
len(merged_second_condensed_merge_allele_list)

merged_third_condensed_merge_allele_list=[]

for sub_sub_merged_allel in merged_second_condensed_merge_allele_list:
    n="".join(set(sub_sub_merged_allel))
    #print(k)
    merged_third_condensed_merge_allele_list.append(n)

#merged_third_condensed_merge_allele_list
len(merged_third_condensed_merge_allele_list)

Counter([len(i) for i in merged_third_condensed_merge_allele_list])
