==> 2.0_DATASET_masters_hapmap_for_tassel_conversion_adding_alleles.ipynb <==
## **2.0_DATASET_masters_hapmap_for_tassel_conversion_adding_alleles.ipynb**

### GOALS of this script:
* merge all filtered into one 
    *intermediate output file 
* edit merged file to fit the hapmap format 
* OUT: hapmap format file
* TASSEL COMMAND: command line tassel command 
* *with prints at different steps foe when the script is run from the command line*
* *checking the genotype calls in the table is currently commented out because otherwise the script dosent run*
* **this scripts edits NA into header columns, NN for the genotype calls and adds the values in the allele column back in to solve the beforehand problem with the wrong encoding of the NN as a second alt allele - see vcf_missing_sites_problem folder**

### *Import packages*

import csv
import pandas as pd
from itertools import chain
import numpy as np
from collections import Counter
import os 

### *Import files* / **2.0 import the files and replace the IUPAC nucleotide code**

**Unterseer_2016_landraces**

for_vcf_filtered_unterseer_2016_landraces_right_pos=pd.read_csv('filtered_unterseer_2016_landraces_for_masters.csv', sep="\t")
#for_vcf_filtered_unterseer_2016_landraces_right_pos.csv

for_vcf_filtered_unterseer_2016_landraces_right_pos

for_vcf_filtered_unterseer_2016_landraces_right_pos=for_vcf_filtered_unterseer_2016_landraces_right_pos.drop(columns=['strand_2016_landraces'])

for_vcf_filtered_unterseer_2016_landraces_right_pos

*Replace calls in the Unterseer files*

for_vcf_filtered_unterseer_2016_landraces_right_pos=for_vcf_filtered_unterseer_2016_landraces_right_pos.replace("A", "AA")
for_vcf_filtered_unterseer_2016_landraces_right_pos=for_vcf_filtered_unterseer_2016_landraces_right_pos.replace("C", "CC")
for_vcf_filtered_unterseer_2016_landraces_right_pos=for_vcf_filtered_unterseer_2016_landraces_right_pos.replace("G", "GG")
for_vcf_filtered_unterseer_2016_landraces_right_pos=for_vcf_filtered_unterseer_2016_landraces_right_pos.replace("T", "TT")
for_vcf_filtered_unterseer_2016_landraces_right_pos=for_vcf_filtered_unterseer_2016_landraces_right_pos.replace("R", "AG")
for_vcf_filtered_unterseer_2016_landraces_right_pos=for_vcf_filtered_unterseer_2016_landraces_right_pos.replace("Y", "CT")
for_vcf_filtered_unterseer_2016_landraces_right_pos=for_vcf_filtered_unterseer_2016_landraces_right_pos.replace("S", "GC")
for_vcf_filtered_unterseer_2016_landraces_right_pos=for_vcf_filtered_unterseer_2016_landraces_right_pos.replace("W", "AT")
for_vcf_filtered_unterseer_2016_landraces_right_pos=for_vcf_filtered_unterseer_2016_landraces_right_pos.replace("K", "GT")
for_vcf_filtered_unterseer_2016_landraces_right_pos=for_vcf_filtered_unterseer_2016_landraces_right_pos.replace("M", "AC")

#for_vcf_filtered_unterseer_2016_landraces_right_pos

**Unterseer_2016_elite**

for_vcf_filtered_unterseer_2016_elite_right_pos=pd.read_csv('filtered_unterseer_2016_elite_for_masters.csv', sep="\t")

for_vcf_filtered_unterseer_2016_elite_right_pos

for_vcf_filtered_unterseer_2016_elite_right_pos=for_vcf_filtered_unterseer_2016_elite_right_pos.drop(columns=['strand_2016_elite'])

for_vcf_filtered_unterseer_2016_elite_right_pos

*Replace calls in the Unterseer files*

for_vcf_filtered_unterseer_2016_elite_right_pos=for_vcf_filtered_unterseer_2016_elite_right_pos.replace("A", "AA")
for_vcf_filtered_unterseer_2016_elite_right_pos=for_vcf_filtered_unterseer_2016_elite_right_pos.replace("C", "CC")
for_vcf_filtered_unterseer_2016_elite_right_pos=for_vcf_filtered_unterseer_2016_elite_right_pos.replace("G", "GG")
for_vcf_filtered_unterseer_2016_elite_right_pos=for_vcf_filtered_unterseer_2016_elite_right_pos.replace("T", "TT")
for_vcf_filtered_unterseer_2016_elite_right_pos=for_vcf_filtered_unterseer_2016_elite_right_pos.replace("R", "AG")
for_vcf_filtered_unterseer_2016_elite_right_pos=for_vcf_filtered_unterseer_2016_elite_right_pos.replace("Y", "CT")
for_vcf_filtered_unterseer_2016_elite_right_pos=for_vcf_filtered_unterseer_2016_elite_right_pos.replace("S", "GC")
for_vcf_filtered_unterseer_2016_elite_right_pos=for_vcf_filtered_unterseer_2016_elite_right_pos.replace("W", "AT")
for_vcf_filtered_unterseer_2016_elite_right_pos=for_vcf_filtered_unterseer_2016_elite_right_pos.replace("K", "GT")
for_vcf_filtered_unterseer_2016_elite_right_pos=for_vcf_filtered_unterseer_2016_elite_right_pos.replace("M", "AC")

#for_vcf_filtered_unterseer_2016_elite_right_pos

**Mayer_2020_dhlines**

for_vcf_filtered_mayer_2020_dhlines_suffix=pd.read_csv('filtered_mayer_2020_dhlines_for_masters.csv', sep="\t")

#for_vcf_filtered_mayer_2020_dhlines_suffix

#for_vcf_filtered_mayer_2020_dhlines_suffix=for_vcf_filtered_mayer_2020_dhlines_suffix.drop(columns=['Unnamed: 0'])

for_vcf_filtered_mayer_2020_dhlines_suffix

**Mayer_2017_landraces**

for_vcf_filtered_mayer_2017_landraces_suffix=pd.read_csv('filtered_mayer_2017_landraces_for_masters.csv', sep="\t")

#for_vcf_filtered_mayer_2017_landraces_suffix

#for_vcf_filtered_mayer_2017_landraces_suffix=for_vcf_filtered_mayer_2017_landraces_suffix.drop(columns=['Unnamed: 0','chr_v2_2017_landraces','pos_v2_2017_landraces'])

for_vcf_filtered_mayer_2017_landraces_suffix

### **2.1 Drop all not needed columns**

**Unterseer_2016_landraces**

#for_vcf_filtered_unterseer_2016_landraces_right_pos

small_for_vcf_filtered_unterseer_2016_landraces_right_pos=for_vcf_filtered_unterseer_2016_landraces_right_pos.drop(columns=['alleles_2016_landraces','variants/CMT_2016_landraces'])
#small_for_vcf_filtered_unterseer_2016_landraces_right_pos

**Unterseer_2016_elite**

#for_vcf_filtered_unterseer_2016_elite_right_pos

small_for_vcf_filtered_unterseer_2016_elite_right_pos=for_vcf_filtered_unterseer_2016_elite_right_pos.drop(columns=['alleles_2016_elite','variants/CMT_2016_elite'])
#small_for_vcf_filtered_unterseer_2016_elite_right_pos

**Mayer_2017_landraces**

#for_vcf_filtered_mayer_2017_landraces_suffix

small_for_vcf_filtered_mayer_2017_landraces_suffix=for_vcf_filtered_mayer_2017_landraces_suffix.drop(columns=['quality_2017_landraces'])
#small_for_vcf_filtered_mayer_2017_landraces_suffix

**Mayer_2020_dhlines**

#for_vcf_filtered_mayer_2020_dhlines_suffix

small_for_vcf_filtered_mayer_2020_dhlines_suffix=for_vcf_filtered_mayer_2020_dhlines_suffix.drop(columns=['quality_2020_dh'])
#small_for_vcf_filtered_mayer_2020_dhlines_suffix

print("Finished - Drop all not needed columns")

### **2.2 Rename the remaining columns**

**Unterseer_2016_elite**

small_for_vcf_filtered_unterseer_2016_elite_right_pos=small_for_vcf_filtered_unterseer_2016_elite_right_pos.rename(columns={"variants/AD_2016_elite": "rs#", "chr_v4_2017_landraces": "chrom","pos_v4_2017_landraces":"pos"})
#small_for_vcf_filtered_unterseer_2016_elite_right_pos

**Unterseer_2016_landraces**

small_for_vcf_filtered_unterseer_2016_landraces_right_pos=small_for_vcf_filtered_unterseer_2016_landraces_right_pos.rename(columns={"variants/AD_2016_landraces": "rs#", "chr_v4_2017_landraces": "chrom","pos_v4_2017_landraces":"pos"})
#small_for_vcf_filtered_unterseer_2016_landraces_right_pos

**Mayer_2020_dhlines**

small_for_vcf_filtered_mayer_2020_dhlines_suffix=small_for_vcf_filtered_mayer_2020_dhlines_suffix.rename(columns={"marker_2020_dh": "rs#", "chr_2020_dh": "chrom","pos_2020_dh":"pos"})
#small_for_vcf_filtered_mayer_2020_dhlines_suffix

**Mayer_2017_landraces**

small_for_vcf_filtered_mayer_2017_landraces_suffix=small_for_vcf_filtered_mayer_2017_landraces_suffix.rename(columns={"index_2017_landraces": "rs#", "chr_v4_2017_landraces": "chrom","pos_v4_2017_landraces":"pos"})
#small_for_vcf_filtered_mayer_2017_landraces_suffix

print("Finished - Rename the remaining columns")

### **2.3 Merge between Unterseer**

merge_marker_unterseer=pd.merge(small_for_vcf_filtered_unterseer_2016_elite_right_pos,small_for_vcf_filtered_unterseer_2016_landraces_right_pos, on=['rs#','chrom','pos'], how='outer')

#merge_marker_unterseer

print("Finished - Merge between Unterseer")

### **2.4 Merge between Mayer**

merge_marker_mayer=pd.merge(small_for_vcf_filtered_mayer_2020_dhlines_suffix,small_for_vcf_filtered_mayer_2017_landraces_suffix,on=['rs#','chrom','pos'], how='outer')

#merge_marker_mayer

print("Finished - Merge between Mayer")

### **2.5 Final merge**

merge_merge=pd.merge(merge_marker_unterseer,merge_marker_mayer,on=['rs#','chrom','pos'], how='outer')

#merge_merge

print("Finished - Final merge")

## **Intermediate OUTPUT I**

### **Write to file**

#merge_merge.to_csv("merge_merge_all_chromosomes_no_edits_for_masters.hmp.txt", sep="\t", index=False)

### **Read in file again**

merge_merge=pd.read_csv("merge_merge_all_chromosomes_no_edits_for_masters.hmp.txt", sep="\t")

merge_merge

### **2.6 Add columns for hapmap format**

#generate new columns
alleles = pd.DataFrame({'alleles' : []})
strand = pd.DataFrame({'strand' : []})
assembly = pd.DataFrame({'assembly' : []})
center = pd.DataFrame({'center' : []})
protLSID = pd.DataFrame({'protLSID' : []})
assayLSID = pd.DataFrame({'assayLSID' : []})
paneLSID = pd.DataFrame({'paneLSID' : []})
QCcode = pd.DataFrame({'QCcode' : []})

#add columns in
merge_merge.insert(loc=1, column = 'alleles', value = alleles)
merge_merge.insert(loc=4, column = 'strand', value = strand)
merge_merge.insert(loc=5, column = 'assembly#', value = assembly)
merge_merge.insert(loc=6, column = 'center', value = center)
merge_merge.insert(loc=7, column = 'protLSID', value = protLSID)
merge_merge.insert(loc=8, column = 'assayLSID', value = assayLSID)
merge_merge.insert(loc=9, column = 'panel', value = paneLSID)
merge_merge.insert(loc=10, column = 'QCcode', value = QCcode)

merge_merge

### **2.7 Additional edits**

**Check the contents of the genotype columns**

#subset the table
check_genotype_counts=merge_merge.iloc[:,11:]

#print(check_genotype_counts)

#get 2D list from the genotype columns
check_genotype_counts_values=check_genotype_counts.values.tolist()

#flatten the list 
flatten_list = list(chain.from_iterable(check_genotype_counts_values))

#get all unique values
flatten_list_unique_values=pd.unique(flatten_list)

print(flatten_list_unique_values)

**2.7.1 Replace NaN with NN**

#replace NaN with NN
merge_merge=merge_merge.fillna("NN")

**2.7.2 Replace of target variant genotype calls OO**

merge_merge=merge_merge.replace(to_replace='OO', value="NN") 

**2.7.3 Replace of N genotype call with NN**

merge_merge=merge_merge.replace(to_replace='N', value="NN") 

#recheck the genotype calls 
#check_genotype_counts_re=merge_merge.iloc[:,11:]
#check_genotype_counts_values_re=check_genotype_counts_re.values.tolist()
#flatten_list_re = list(chain.from_iterable(check_genotype_counts_values_re))
#flatten_list_unique_values_re=pd.unique(flatten_list_re)

#print(flatten_list_unique_values_re)
len(flatten_list_unique_values_re)

**2.7.4 Turn around double calls**

*Exchange C/T for T/C*

merge_merge=merge_merge.replace(to_replace='CT',value='TC')

*Exchange G/T for T/G*

merge_merge=merge_merge.replace(to_replace='GT',value='TG')

*Exchange A/T for T/A*

merge_merge=merge_merge.replace(to_replace='AT',value='TA')

*Exchange A/G for G/A*

merge_merge=merge_merge.replace(to_replace='AG',value='GA')

*Exchange C/G for G/C*

merge_merge=merge_merge.replace(to_replace='CG',value='GC')

*Exchange C/A for A/C*

merge_merge=merge_merge.replace(to_replace='CA',value='AC')

merge_merge

## **Intermediate OUTPUT II**

### **Write to file**

merge_merge.to_csv("merge_merge_all_chromosomes_turn_around_for_masters.hmp.txt", sep="\t", index=False)

### **Read in file again**

merge_merge_turn_around=pd.read_csv("merge_merge_all_chromosomes_turn_around_for_masters.hmp.txt", sep="\t")

merge_merge_turn_around

#recheck the genotype calls 
check_genotype_counts_after_turn=merge_merge_turn_around.iloc[:,11:]

check_genotype_counts_values_after_turn=check_genotype_counts_after_turn.values.tolist()

#flatten_list_after_turn= list(chain.from_iterable(check_genotype_counts_values_after_turn))

#flatten_list_unique_values_after_turn=pd.unique(flatten_list_after_turn)

#flatten_list_unique_values_after_turn
#len(flatten_list_unique_values_after_turn)

**2.7.5 Replace the header columns with NA**

#select the first 11 columns 
first_11=merge_merge_turn_around.iloc[:,:11]

first_11

#replace the header NN with with a NA 
merge_merge_turn_around.iloc[:,0:11]=merge_merge_turn_around.iloc[:,0:11].replace(to_replace='NN',value="NA")

#select the first 11 columns again
first_11_2=merge_merge_turn_around.iloc[:,:11]

first_11_2

**2.7.6 Make allele column for table**

*Get the row-wise alleles from the table*

merge_merge_set=list(map(set,check_genotype_counts_after_turn.values))

#list of sets 
merge_merge_set

len(merge_merge_set)

*Make list of sets into list of lists*

#mayer_2017_genotypes
merge_merge_list=[]

for e in merge_merge_set:
    #print(e)
    geno_row=list(e)
    merge_merge_list.append(geno_row)

merge_merge_list

len(merge_merge_list)

#check the genotype calls that are contained
flatten_list= list(chain.from_iterable(merge_merge_list))
flatten_list_unique_values_merge_merge_list=pd.unique(flatten_list)

flatten_list_unique_values_merge_merge_list
len(flatten_list_unique_values_merge_merge_list)

print(flatten_list_unique_values_merge_merge_list)

*Remove "NN" from the rows*

for sublist in merge_merge_list:
    #print(sublist)
    for element in sublist:
        #print(element)
        if element == "NN":
            #print("yes")
            sublist.remove(element)

merge_merge_list

len(merge_merge_list)

*Merge all the row elements into one*

merge_merge_genotypes_list=[]

for sublist_without_NN in merge_merge_list:
    #print(sublist_without_2017)
    s=''.join(sublist_without_NN)
    merge_merge_genotypes_list.append(s)
    #print(s)

merge_merge_genotypes_list

len(merge_merge_genotypes_list)

substring_T="T"
substring_G="G"
substring_A="A"
substring_C="C"

count_row=[]
all_counts=[]
filter_boolean=[]

*Check for sites with more than two alleles*

for merged_element in merge_merge_genotypes_list:
    #print(merged_element)
    
    #count how many times each allel appears in the row 
    count_T=merged_element.count(substring_T)
    count_G=merged_element.count(substring_G)
    count_A=merged_element.count(substring_A)
    count_C=merged_element.count(substring_C)
    #print(count_T)
    
    #add the rowwise counts of the alleles to a list
    count_row=[count_T,count_G,count_A,count_C]
    #add the row counts to a list of lists
    all_counts.append(count_row)

    #count number of non-zero numbers in the count of alleles list (are there more than two alleles )
    num_of_alleles=np.count_nonzero(count_row)
    #print(num_of_alleles)
    
    if num_of_alleles > 2: #if there are more than two alleles 
        #print(merged_element)
        #print(num_of_alleles)
        filter_boolean.append(False) #append False to the list 
    
    elif num_of_alleles <=2: #if there are less than two allels
        filter_boolean.append(True)
              
    else:
        print("Problem")
        break  

print(len(filter_boolean))
Counter(filter_boolean)

*Make the allele column*

allel_column=[]

#make the allel column 
for count_row_element in all_counts :
            #print(count_row_element)
            if count_row_element[0] != 0 and count_row_element[1] != 0:
                allel_column.extend(["T/G"])
            elif count_row_element[0] != 0 and count_row_element[2] !=0:
                allel_column.extend(["T/A"])
            elif count_row_element[0] != 0 and count_row_element[3] !=0:
                allel_column.extend(["T/C"])
            elif count_row_element[1] != 0 and count_row_element[2] !=0:
                allel_column.extend(["G/A"])
            elif count_row_element[1] != 0 and count_row_element[3] !=0:
                allel_column.extend(["G/C"])
            elif count_row_element[2] != 0 and count_row_element[3] !=0:
                allel_column.extend(["A/C"])
            elif count_row_element[0] != 0 and count_row_element[1] ==0 and count_row_element[2] ==0 and count_row_element[3]==0:
                allel_column.extend(["T"])
            elif count_row_element[1] != 0 and count_row_element[0] ==0 and count_row_element[2] ==0 and count_row_element[3]==0:
                allel_column.extend(["G"])
            elif count_row_element[2] != 0 and count_row_element[0] ==0 and count_row_element[1] ==0 and count_row_element[3]==0:
                allel_column.extend(["A"])
            elif count_row_element[3] != 0 and count_row_element[0] ==0 and count_row_element[1] ==0 and count_row_element[2]==0:
                allel_column.extend(["C"])
            else:
                allel_column.extend(["-/-"])


len(allel_column)

Counter(allel_column)

205756+203870+49079+35345+57468+57654+56+59+66+89

56+59+66+89

allel_column

*Add the allele column into the table*

merge_merge_turn_around

#read-in the array.txt
check_strand=pd.read_csv("/Users/kschul38/Documents/tsinfer-project/data/2_processed/europe_maize_dataset_600k/600k_array_info_header_removed.txt", sep="\t", dtype = str)
check_strand

#get all the genotype-calls from the array.txt
genotype_call=check_strand["alleles_f"].values.ravel()

#check that number of genotype calls is correct
len(genotype_call)

#get a list of all unique calls
unique_values_array = pd.unique(genotype_call)

#look at complete list
#unique_values_array

#remove the "typical genotype calls" - so that just indels remain
unique_values_array = np.delete(unique_values_array, [0,1,2,3,4,5])

#check that it only contains a-typical genotype calls
#unique_values_array

#create a boolean over the whole allele column, where typical calls == True & atypical calls == False 
array_marker_filter_table=~check_strand["alleles_f"].isin(unique_values_array)

array_marker_filter_table

**Get the filtered markers from the array.txt**

filter_boolean_array_txt=array_marker_filter_table #taken from the non-typical genotypes removed to use here
#Counter(filter_boolean_array_txt)

#filter_boolean_array_txt

Counter(filter_boolean_array_txt)

#check_strand[["ID","alleles_f"]]
array_txt_alleles=check_strand[["ID","alleles_f"]]
array_txt_alleles_merge=array_txt_alleles.rename(columns={"ID":"rs#"}) #rename the column for merge 
array_txt_alleles_merge_filter=array_txt_alleles_merge[filter_boolean_array_txt] # remove the lines with ne non-typical genotype calls 

array_txt_alleles_merge_filter

merge_merge_with_column=pd.merge(array_txt_alleles_merge_filter,merge_merge_turn_around,on=['rs#'], how='outer')

merge_merge_with_column

merge_merge_with_column = merge_merge_turn_around.assign(alleles=allel_column)

merge_merge_with_column

col_check=merge_merge_with_column.loc[merge_merge_with_column['rs#']=="AX-91835112"]
col_check

pd.unique(col_check.iloc[:,11:].values.ravel())

**2.7.7 Reorder first by chromosome than by position**

#sort by chromosome and then by position
merge_merge_with_column=merge_merge_with_column.sort_values(by=['chrom', 'pos'])

**2.7.8 Convert the position of the markers into integer**

merge_merge_with_column[["pos"]]=merge_merge_with_column[["pos"]].astype(int)

merge_merge_with_column

#replace the header NN with with a NA 
merge_merge_with_column.iloc[:,0:11]=merge_merge_with_column.iloc[:,0:11].replace(to_replace='NN',value="NA")

print("Finished - Additional edtis")

merge_merge_with_column

merge_merge_with_column=merge_merge_with_column.drop(columns=["alleles"])

merge_merge_with_column

merge_merge_with_column=merge_merge_with_column.rename(columns={"alleles_f":"alleles"})

merge_merge_with_column

## **OUTPUT**

### **Write to file**

merge_merge_with_column.to_csv("merge_merge_all_chrom_all_edits_with_allel_column_from_array_600k_for_masters.hmp.txt", sep="\t", index=False)

print("Finished written to file")

### **TASSEL command**

%cd /Users/kschul38/Documents/Tassel_5.0

os.system("./run_pipeline.pl -Xmx10g  -h merge_merge_all_chrom_all_edits_with_allel_column_from_array_600k_for_masters.hmp.txt -export merge_merge_all_chrom_all_edits_with_allel_column_from_array_600k_for_masters_vcf.vcf -exportType VCF")
